+++ Sources/BatteryPanicApp/Shared/AppConstants.swift
@@ -5,5 +5,4 @@ enum AppConstants {
     static let defaultThreshold = 10
     static let pollInterval: TimeInterval = 5
     static let previewAlarmDuration: TimeInterval = 4
-    static let warningSoundRepeatInterval: TimeInterval = 10
 }